Backend/python_api/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-multipart
pymysql
cryptography
mysql-connector-python
python-jose[cryptography]
passlib[bcrypt]